+++ tests/test_plotly_figures/test_plotly_scattermapbox_figures_project_page.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import pytest
 from plotly.graph_objs import Figure
 
-from src.linear_objects.project import get_projects_from_saved_data
+from src.linear_objects.reinforcement_program import get_projects_from_saved_data
 from src.plotly_graphs.project_page.plotly_maps import plot_project_overview_map